server/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
httpx
ollama
python-jose[cryptography]
passlib[argon2]
google-auth
requests
sqlalchemy
pymysql
python-dotenv
psycopg2-binary
alembic
